requirements.txt
pandas==1.3.3
openpyxl==3.0.9
mysql-connector-python==8.0.26